+++ src/anyio/_backends/_trio.py
@@ -3,7 +3,7 @@ from types import TracebackType
 from typing import Callable, Optional, List, Type, Union
 
 import trio.from_thread
-from async_generator import async_generator, yield_, asynccontextmanager, aclosing
+from async_generator import async_generator, yield_, asynccontextmanager
 from trio.to_thread import run_sync
 
 from .. import abc, claim_worker_thread, T_Retval, TaskInfo
@@ -31,7 +31,6 @@ run = trio.run
 # Miscellaneous
 #
 
-finalize = aclosing
 sleep = trio.sleep