+++ tests/conftest.py
@@ -15,10 +15,6 @@ else:
             pytest.mark.skip(reason='uvloop is missing shutdown_default_executor()'))
 
 
-def pytest_ignore_collect(path, config):
-    return path.basename.endswith('_py36.py') and sys.version_info < (3, 6)
-
-
 @pytest.fixture(params=[
     pytest.param(('asyncio', {'use_uvloop': False}), id='asyncio'),
     pytest.param(('asyncio', {'use_uvloop': True}), id='asyncio+uvloop', marks=uvloop_marks),