+++ docs/versionhistory.rst
@@ -5,6 +5,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 **UNRELEASED**
 
+- **BACKWARDS INCOMPATIBLE** Removed the ``anyio.finalize()`` context manager since as of curio
+  1.0, it is no longer necessary. Use ``async_generator.aclosing()`` instead.
 - Bumped minimum versions of trio and curio to v0.16 and v1.2, respectively
 
 **1.4.0**